Extract Product data class
