+++ androidApp/src/main/java/com/thinlineit/favorit_android/android/data/entity/Funding.kt
@@ -23,9 +23,3 @@ data class Funding(
 enum class FundingState {
     OPENED, EXPIRED, CLOSED, COMPLETED
 }
-
-data class Product(
-    val link: String,
-    val option: String,
-    val price: Int,
-)
