+++ androidApp/src/main/java/com/thinlineit/favorit_android/android/data/entity/Product.kt
@@ -1,0 +1,7 @@
+package com.thinlineit.favorit_android.android.data.entity
+
+data class Product(
+    val link: String,
+    val option: String,
+    val price: Int,
+)
