+++ androidApp/src/main/java/com/thinlineit/favorit_android/android/ui/createfunding/usecase/CreateFunding.kt
@@ -1,7 +1,7 @@
 package com.thinlineit.favorit_android.android.ui.createfunding.usecase
 
 import com.thinlineit.favorit_android.android.data.Result
-import com.thinlineit.favorit_android.android.data.repository.CreateFundingRequest
+import com.thinlineit.favorit_android.android.data.entity.CreateFundingRequest
 import com.thinlineit.favorit_android.android.data.repository.FundingRepository
 import com.thinlineit.favorit_android.android.ui.createfunding.CreateFundingResult
